bridgeit-motion-simulator updates
Fixed a bug where the correct location was not being used when it is
selected in the autocomplete. Centre the map on the origin and
destination when a location is selected from the autocomplete. Add a
toggle for “follow user” when running a motion simulation. Don't allow
map refreshes during simulation. Doc updates.