location-simulator updates
Removed follow user option for now. Instead when playAll is fired from
the simulator we will create a new bounds object that is only meant for
that particular simulation. The routes can use this to keep the map
centred and panned on the simulation. This solution still needs some
work because it doesn’t work well with individual routes.